requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
groq==0.11.0
python-dotenv==1.0.1
supabase==1.0.1
gunicorn==23.0.0
